% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{prepare_long_format}
\alias{prepare_long_format}
\title{prepare_long_format}
\usage{
prepare_long_format(data_list, data_list_names)
}
\arguments{
\item{data_list}{a csv data file in the wide format}

\item{data_list_names}{a vector of study names}
}
\value{
a data frame in the long format
}
\description{
prepare_long_format
}
